# openapi.yaml
openapi: 3.0.0
info:
  title: task-api
  description: >-
    Task management REST API. Users own projects; projects contain tasks; access
    is scoped to the authenticated user at every layer.
  version: 0.1.0
...
paths:
...
  /auth/register:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        '201':
          description: A user
          content:
            application/json:
              schema:
                description: A user
                type: object
                required:
                  - id
                  - email
                  - displayName
                  - createdAt
                properties:
                  id:
                    type: integer
                  email:
                    type: string
                  displayName:
                    type: string
                  createdAt:
                    type: string
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
      responses:
        '200':
          description: An access/refresh token pair
          content:
            application/json:
              schema:
                description: An access/refresh token pair
                type: object
                required:
                  - accessToken
                  - refreshToken
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
  /auth/refresh:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              additionalProperties: false
              properties:
                refreshToken:
                  type: string
                  pattern: ^[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+$
                  maxLength: 2048
      responses:
        '200':
          description: An access/refresh token pair
          content:
            application/json:
              schema:
                description: An access/refresh token pair
                type: object
                required:
                  - accessToken
                  - refreshToken
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
...